backend/requirements.txt
Flask-Migrate
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
bcrypt
cryptography
# Add other dependencies as needed

# For Auth0 JWT validation (uncomment when ready)
# PyJWT
# cryptography
# requests